update imgui to be able to edit important value

+++ KrillKrew/KrillEngine/KrillEngine/ProjectileObject.cpp
@@ -153,7 +153,7 @@ void ProjectileObject::ExplodeTileInRange()
 	this->isActive = false;
 }
 
-int ProjectileObject::GetLifetime() 
+float ProjectileObject::GetLifetime()
 {
 	return lifeTime;
 }
